
\name{DeducerMMR}
\title{
 A Deducer plugin for moderated multiple regressions and simple slopes analysis
}
\description{
Using pequod package, DeducerMMR performs a) moderated multiple regressions with both uncentered and centered predictors, and with residual centering, b) simple slopes analysis and c) a slopes plot. 
}

\references{
Little, T. D., Bovaird, J. A., & Widaman, K. F. (2006). On the Merits of Orthogonalizing Powered and Product Terms: Implications for Modeling Interactions Among Latent Variables. Structural Equation Modeling, 13(4), 497-519.


Cohen, J., Cohen, P.,West, S. G.,&Aiken, L. S. (2003). Applied multiple regression/correlation analysis for the behavioral sciences (3rd ed.). Mahwah, NJ: Lawrence Erlbaum Associates, Inc.


Bauer, D. J., & Curran, P. J. (2005). Probing Interactions in Fixed and Multilevel Regression: Inferential and Graphical Techniques. Multivariate Behavioral Research, 40(3), 373-400.


Dawson, J. F., and Richter, W. (2006) Probing Three-Way interactions in moderated multiple regression: Development and application of a slope difference test. Journal of Applied Psychology, 4, 917-926.

}