\name{numSelected}
\alias{numSelected}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extracting numeric variables
}
\description{
a function to extract numeric variables of a lmres model starting from the string vector of the coded variables.
}
\usage{
numSelected(mod1, code)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mod1}{
a lmres object
}
  \item{code}{
a string vector of coded variables
}
}

\value{
numSelected returns a string vector of numeric variables
}




% Add one or more standard keywords, see file 'KEYWORDS' in 
% the R documentation directory.
\keyword{ ~kwd1 }

